Compare InChI assignment 1.03/1.06
Since today (2021-09-25), Debichem offers an InChI executable in branch testing
for Debian 12/bookworm.  Possibly for consistency with the other libraries of inchi
(1.03+dfsg-4), this inchi_main however refers to the old version 1.03 by June 15,
2010, limited to .mol/.sdf files of format v2000.

The 5k set was scrutinized, in the elder v2000 as well as the v3000 format, for an
InChI assignment with this executable.  The assignments were compared with the
executable currently distributed by the InChI trust (1.06, December 2020) on both
the old as well as the new format.

No differences for processing the old format.  6/5k differences for comparing the
assignment by the old/the new executable (again, possibly the motif of oxime/imidic acid).